+++ apps/aspiration/serializers/aspirations.py
@@ -1,29 +1,6 @@
 from rest_framework import serializers
-from .models import Category, Aspiration, AspirationProgress, Notification
-
-class CategorySerializer(serializers.ModelSerializer):
-    name = serializers.CharField(
-        required=True,
-        error_messages={
-            'required': 'Nama kategori wajib diisi.',
-            'blank': 'Nama kategori tidak boleh kosong.'
-        }
-    )
-    color = serializers.SerializerMethodField()
-
-    class Meta:
-        model = Category
-        fields = ['id', 'name', 'description', 'color']
-
-    def get_color(self, obj):
-        mapping = {
-            'fasilitas': '#6D5DFF',
-            'lingkungan': '#9DC344',
-            'pendidikan': '#C03648',
-            'karakter': '#C08736',
-            'ibadah': '#BA36C0',
-        }
-        return mapping.get(obj.name.lower(), '#6C757D')
+from apps.aspiration.serializers.categories import CategorySerializer
+from apps.aspiration.models import Category, Aspiration, AspirationProgress
 
 
 class AspirationProgressSerializer(serializers.ModelSerializer):
@@ -122,37 +99,6 @@ class AspirationSerializer(serializers.ModelSerializer):
                 "nis": student.nis,
                 "rombel": student.rombel,
                 "rayon": student.rayon,
-                # "image": student.image.url if student.image else None
-            }
-        except:
-            return None
-
-
-class NotificationSerializer(serializers.ModelSerializer):
-    report_id = serializers.CharField(source='aspiration.report_id', read_only=True)
-    aspiration_title = serializers.CharField(source='aspiration.title', read_only=True)
-    aspiration_description = serializers.CharField(source='aspiration.description', read_only=True)
-    
-    student_info = serializers.SerializerMethodField()
-    student = serializers.CharField(source='aspiration.user.email', read_only=True)
-    
-    class Meta:
-        model = Notification
-        fields = [
-            'id',  'report_id', 'message', 'student', 'student_info',
-            'aspiration_title', 'aspiration_description', 
-            'is_read', 'created_at'
-        ]
-        read_only_fields = ['id', 'report_id', 'message', 'created_at']
-
-    def get_student_info(self, obj):
-        try:
-            student_profile = obj.aspiration.user.student_profile 
-            return {
-                "name": student_profile.name,
-                "nis": student_profile.nis,
-                "rombel": student_profile.rombel,
-                "rayon": student_profile.rayon,
                 # "image": student.image.url if student.image else None
             }
         except: